Split event and state for `ChooseSelfVerificationModePresenter` into their own files
